requirements.txt
aiohttp==3.7.4.post0
anyio==3.5.0
asgiref==3.5.0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
click==8.0.4
discord.py==1.7.3
environs==9.5.0
fastapi==0.75.0
h11==0.13.0
idna==3.3
marshmallow==3.15.0
multidict==6.0.2
packaging==21.3
pydantic==1.9.0
pyparsing==3.0.7
python-dotenv==0.19.2
quantumrandom==1.9.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
yarl==1.7.2
